Реализована простейшая работа с СУБД (выборка и вставка)

+++ src/mebel/sqldatabasehelper.cpp
@@ -3,13 +3,15 @@
 #include <QSqlQuery>
 #include <QDebug>
 #include <QSqlError>
+#include <QSqlRecord>
 
 SqlDatabaseHelper::SqlDatabaseHelper()
     : isOpen_(false)
 {
 }
 
-void SqlDatabaseHelper::open(const QString &type, const QString &hostName,
+void SqlDatabaseHelper::open(const QString &type,
+                             const QString &hostName,
                              const QString &databaseName,
                              const QString &userName, const QString &password)
 {
@@ -17,51 +19,62 @@ void SqlDatabaseHelper::open(const QString &type, const QString &hostName,
     db_.setHostName(hostName);
     db_.setDatabaseName(databaseName);
     isOpen_ = db_.open(userName, password);
-    qDebug() << db_.lastError().text();
-    qDebug() << db_.lastError().databaseText();
-    qDebug() << db_.lastError().driverText();
+    if (!isOpen_) {
+        qDebug() << "text" << db_.lastError().text();
+        qDebug() << "databaseText" << db_.lastError().databaseText();
+        qDebug() << "driverText" << db_.lastError().driverText();
+    }
 }
 
 void SqlDatabaseHelper::close()
 {
+    if (!isOpen_) {
+        // throw
+        return;
+    }
     QSqlDatabase::removeDatabase(db_.connectionName());
 }
 
 QStringList SqlDatabaseHelper::select(const QString &tableName,
                                       const QString &filter)
 {
-    qDebug() << isOpen_;
     if (!isOpen_) {
 //        throw
         return QStringList();
     }
 
-
     QStringList values;
-    QSqlQuery query(QString("SELECT * FROM %0").arg(tableName), db_);
+
+    QSqlQuery query(QString("SELECT * FROM %0;").arg(tableName), db_);
+    int columnCount = query.record().count();
+
     while (query.next()) {
-        qDebug() << query.value(0);
+        for (int i = 0; i < columnCount; ++i) {
+            values << query.value(i).toString();
+        }
     }
     return values;
 }
 
-void SqlDatabaseHelper::insert(const QString &tableName,
+quint32 SqlDatabaseHelper::insert(const QString &tableName,
                                const QStringList &columns,
                                const QStringList &values)
 {
     if (!isOpen_) {
 //        throw
-        return;
+        return 0;
     }
 
-    QString columnsStr = columns.join(", ");
-    QString valuesStr = values.join(", ");
-    QString queryStr = QString("INSERT INTO %1 (%2) VALUES (%3) RETURNING id").arg(tableName)
-            .arg(columnsStr)
-            .arg(valuesStr);
-    qDebug("%s", queryStr.toStdString().c_str());
-    QSqlQuery query(queryStr, db_);
-    while (query.next()) {
-//        id = query.value(0).toInt();
+    QString columnsWithCommas = columns.join(",");
+    QString valuesWithCommas = values.join(",");
+    QSqlQuery query(QString("INSERT INTO %0 (%1) VALUES (%2);").arg(tableName)
+                    .arg(columnsWithCommas).arg(valuesWithCommas), db_);
+    if (query.lastError().type() != QSqlError::NoError) {
+        qDebug() << "text" << query.lastError().text();
+        qDebug() << "databaseText" << query.lastError().databaseText();
+        qDebug() << "driverText" << query.lastError().driverText();
+        return 0;
     }
+
+    return query.lastInsertId().toInt();
 }
